+++ cmd/gaiad/cmd/testnet.go
@@ -30,11 +30,8 @@ import (
 	tmrand "github.com/tendermint/tendermint/libs/rand"
 	"github.com/tendermint/tendermint/types"
 	tmtime "github.com/tendermint/tendermint/types/time"
-
 	// ibcclienttypes "github.com/cosmos/ibc-go/v4/modules/core/02-client/types"
 	// ibcchanneltypes "github.com/cosmos/ibc-go/v4/modules/core/04-channel/types"
-
-	"github.com/cosmos/gaia/v9/app/params"
 )
 
 var (
@@ -149,9 +146,7 @@ func InitTestnet(
 	nodeIDs := make([]string, numValidators)
 	valPubKeys := make([]cryptotypes.PubKey, numValidators)
 
-	simappConfig := params.CustomAppConfig{
-		Config: *srvconfig.DefaultConfig(),
-	}
+	simappConfig := *srvconfig.DefaultConfig()
 	simappConfig.MinGasPrices = minGasPrices
 	simappConfig.API.Enable = true
 	simappConfig.Telemetry.Enabled = true